Refactor catalog dashboards into dedicated pages
